==> notebooks/02_analyze_data.ipynb <==
import re
from collections import defaultdict

import pandas as pd
import sentencepiece as spm
from uralicNLP import uralicApi

data = pd.read_csv("./data/lexeme_stems.csv")

data.info()

data["contlex"].value_counts()

len(data["contlex"].unique())

data[data["contlex"] == "V_LAEULLAD"]

## Prepare data for BPE

data["pos"].unique()

selected_pos_data = data[data['pos'].isin(['N', 'V'])]

pattern = re.compile(r'^[A-Z]+_.+')

filtered_data = selected_pos_data[selected_pos_data['contlex'].apply(lambda x: bool(pattern.match(x)))]

filtered_data

miniparadigms = {
    "V": ["V+Ind+Prs+ConNeg", "V+Ind+Prs+Sg3", "V+Ind+Prt+Sg1", "V+Ind+Prt+Sg3", "V+Inf", "V+Ind+Prs+Sg1", "V+Pass+PrfPrc", "V+Ind+Prs+Pl3", "V+Imprt+Sg3", "V+Imprt+Pl3"],
    "N": ["N+Sg+Loc", "N+Sg+Ill", "N+Pl+Gen", "N+Sg+Nom", "N+Sg+Gen", "N+Sg+Loc+PxSg3", "N+Ess", "N+Der/Dimin+N+Sg+Nom", "N+Der/Dimin+N+Sg+Gen", "N+Sg+Ill+PxSg1"],
    # "A": ["A+Attr"],
}

for index, row in filtered_data.iterrows():
    lexeme = row['lexeme']
    pos = row['pos']
    all_forms = []
    if pos in miniparadigms:
        forms = miniparadigms[pos]
        for form in forms:
            result = uralicApi.generate(f"{lexeme}+{form}", "sms", dictionary_forms=False)
            if result:
                all_forms.extend([_f[0] for _f in result])
        
        filtered_data.loc[index, "forms"] = " ".join([_f for _f in all_forms])

filtered_data.to_csv('./data/lexeme_data_with_forms_3.csv', index=False)

with open('./data/bpe_lexeme_text_3.txt', 'w', encoding='utf-8') as f:
    f.write('\n'.join([f"{row['lexeme']} {row['forms']}" for _, row in filtered_data.iterrows()]))

spm.SentencePieceTrainer.Train('--input=./data/bpe_lexeme_text_3.txt --model_prefix=skolt_bpe_3 --vocab_size=2000 --model_type=bpe')
