==> notebooks/03_train_model_and_test.ipynb <==
import pandas as pd
import torch
import torch.nn as nn
import torch.optim as optim
from torch.utils.data import DataLoader, Dataset, random_split
from torch.cuda.amp import GradScaler, autocast  # For mixed precision training
from sklearn.model_selection import StratifiedShuffleSplit
from sklearn.model_selection import train_test_split
import sentencepiece as spm 
import lightning as L
from lightning.pytorch.tuner import Tuner
from torchmetrics.classification import MulticlassF1Score
from lightning.pytorch.callbacks import ModelCheckpoint, StochasticWeightAveraging
from lightning.pytorch.callbacks.early_stopping import EarlyStopping
import numpy as np
from sklearn.preprocessing import LabelEncoder
from sklearn.metrics import f1_score, accuracy_score, classification_report
import pickle
from datetime import datetime

data = pd.read_csv("./data/lexeme_data_with_forms_3.csv")

data

clean_data = data.copy()
clean_data.dropna(subset=["forms"], inplace=True)
clean_data.drop(columns=["lexeme_and_forms", "bpe"], inplace=True)

clean_data

len(clean_data["contlex"].unique())

clean_data["label"] = clean_data["contlex"].str.split('_').str[:2].str.join('_')

clean_data

label_counts = clean_data["label"].value_counts()
label_counts

counts_of_counts = label_counts.value_counts()
counts_of_counts

def remove_rare_combinations(data, contlex_column='contlex', min_samples=20):
    label_counts = data[contlex_column].value_counts()
    
    frequent_labels = label_counts[label_counts >= min_samples].index

    filtered_data = data[data[contlex_column].isin(frequent_labels)]
    
    return filtered_data.reset_index(drop=True)


clean_data = remove_rare_combinations(clean_data, contlex_column='label', min_samples=50)

clean_data

clean_data["label"].value_counts()

len(clean_data["label"].unique())

clean_data.groupby("pos")["label"].nunique().to_dict()

sp = spm.SentencePieceProcessor(model_file='./skolt_bpe_3.model')

class LabelEncoderManager:
    def __init__(self):
        self.pos_encoder = LabelEncoder()
        self.contlex_encoders = {}  # Dictionary to hold a LabelEncoder for each POS class

    def fit(self, pos_labels, contlex_labels):
        """
        Fit the POS encoder and the contlex encoders based on the corresponding POS.
        
        pos_labels: List or array of POS labels.
        contlex_labels: List or array of contlex labels.
        """
        # Fit the POS encoder
        self.pos_encoder.fit(pos_labels)

        # Initialize and fit a LabelEncoder for each unique POS class
        unique_pos = set(pos_labels)
        for pos_class in unique_pos:
            contlex_for_pos = [contlex_labels[i] for i in range(len(pos_labels)) if pos_labels[i] == pos_class]
            encoder = LabelEncoder()
            encoder.fit(contlex_for_pos)
            self.contlex_encoders[pos_class] = encoder

    def transform_pos(self, pos_labels):
        """Transform POS labels using the POS encoder."""
        return self.pos_encoder.transform(pos_labels)

    def inverse_transform_pos(self, encoded_pos_labels):
        """Inverse transform encoded POS labels using the POS encoder."""
        return self.pos_encoder.inverse_transform(encoded_pos_labels)

    def transform_contlex(self, pos_labels, contlex_labels):
        """
        Transform contlex labels using the corresponding encoder for each POS class.
        
        pos_labels: List or array of POS labels (used to select the corresponding contlex encoder).
        contlex_labels: List or array of contlex labels to be encoded.
        """
        encoded_contlex_labels = []
        for pos, contlex in zip(pos_labels, contlex_labels):
            encoder = self.contlex_encoders.get(pos)
            if encoder is not None:
                encoded_contlex_labels.append(encoder.transform([contlex])[0])
            else:
                raise ValueError(f"No contlex encoder found for POS class: {pos}")
        return encoded_contlex_labels

    def inverse_transform_contlex(self, pos_labels, encoded_contlex_labels):
        """
        Inverse transform contlex labels using the corresponding encoder for each POS class.
        
        pos_labels: List or array of POS labels (used to select the corresponding contlex encoder).
        encoded_contlex_labels: List or array of encoded contlex labels to be inverse transformed.
        """
        contlex_labels = []
        for pos, enc_contlex in zip(pos_labels, encoded_contlex_labels):
            encoder = self.contlex_encoders.get(pos)
            if encoder is not None:
                contlex_labels.append(encoder.inverse_transform([enc_contlex])[0])
            else:
                raise ValueError(f"No contlex encoder found for POS class: {pos}")
        return contlex_labels

    def save_encoders(self, path):
        """
        Save the POS encoder and contlex encoders to a file using pickle.
        
        path: The file path where the encoders will be saved.
        """
        with open(path, 'wb') as f:
            pickle.dump({'pos_encoder': self.pos_encoder, 'contlex_encoders': self.contlex_encoders}, f)

    def load_encoders(self, path):
        """
        Load the POS encoder and contlex encoders from a file using pickle.
        
        path: The file path where the encoders are stored.
        """
        with open(path, 'rb') as f:
            data = pickle.load(f)
            self.pos_encoder = data['pos_encoder']
            self.contlex_encoders = data['contlex_encoders']

device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')

class CustomDataset(Dataset):
    def __init__(self, X, pos_labels, contlex_labels):
        self.X = X
        self.pos_labels = pos_labels
        self.contlex_labels = contlex_labels

    def __len__(self):
        return len(self.X)

    def __getitem__(self, idx):
        X_item = torch.tensor(self.X[idx], dtype=torch.long)
        pos_label = torch.tensor(self.pos_labels[idx], dtype=torch.long)
        contlex_label = torch.tensor(self.contlex_labels[idx], dtype=torch.long)
        return X_item, pos_label, contlex_label

clean_data["X"] = clean_data["lexeme"] + " " + clean_data["forms"]

training_data, test_data = train_test_split(clean_data, test_size=0.1, random_state=42, stratify=clean_data["label"])

training_data

def tokenize_input(texts, sp_model):
    """
    Tokenize input text using a SentencePiece model.
    
    texts: List of input texts to be tokenized.
    sp_model: SentencePiece model to tokenize the texts.
    
    Returns:
    tokenized_texts: List of tokenized and padded input sequences.
    max_len: Maximum sequence length.
    vocab_size: Size of the vocabulary.
    """
    tokenized_texts = [sp_model.encode(text, out_type=int) for text in texts]
    max_len = max(len(x) for x in tokenized_texts)
    padded_texts = [x + [0] * (max_len - len(x)) for x in tokenized_texts]  # Padding with 0
    return padded_texts, max_len, sp_model.get_piece_size()

X, max_len, vocab_size = tokenize_input(training_data['X'].tolist(), sp)
pos_labels = training_data['pos'].tolist()
contlex_labels = training_data['label'].tolist()

encoder_manager = LabelEncoderManager()

encoder_manager.fit(pos_labels, contlex_labels)

encoded_pos = encoder_manager.transform_pos(pos_labels)
encoded_contlex = encoder_manager.transform_contlex(pos_labels, contlex_labels)

# encoder_manager.save_encoders('label_encoders_3.pkl')

encoder_manager.load_encoders('label_encoders_3.pkl')

contlex_output_map = training_data.groupby("pos")["label"].nunique().to_dict()
contlex_output_map = {encoder_manager.transform_pos([k])[0]: v for k, v in contlex_output_map.items()}
contlex_output_map

X_train, X_val, pos_train, pos_val, contlex_train, contlex_val = train_test_split(X, encoded_pos, encoded_contlex, test_size=0.2, random_state=42, stratify=contlex_labels)

train_dataset = CustomDataset(X_train, pos_train, contlex_train)
val_dataset = CustomDataset(X_val, pos_val, contlex_val)

train_dataloader = DataLoader(train_dataset, batch_size=512, shuffle=True)
val_dataloader = DataLoader(val_dataset, batch_size=512, shuffle=False)

class SharedEmbeddingTransformer(L.LightningModule):
    def __init__(
        self,
        pos_num_classes,
        contlex_output_map,
        vocab_size,
        embed_size=96,
        hidden_size=128,
        num_layers=2,
        nhead=4,
        dropout=0.1,
        learning_rate=1e-3,
        batch_size=32,
    ):
        super(SharedEmbeddingTransformer, self).__init__()
        self.save_hyperparameters()

        self.embedding = nn.Embedding(
            num_embeddings=vocab_size, embedding_dim=embed_size, padding_idx=0
        )

        encoder_layer = nn.TransformerEncoderLayer(
            d_model=embed_size,
            nhead=nhead,
            dim_feedforward=hidden_size,
            dropout=dropout,
            batch_first=True
        )
        self.transformer_encoder = nn.TransformerEncoder(encoder_layer, num_layers=num_layers)

        self.fc_out_pos = nn.Linear(embed_size, pos_num_classes)

        self.fc_out_contlex = nn.ModuleDict(
            {
                str(pos_class): nn.Linear(embed_size, contlex_output_map[pos_class])
                for pos_class in contlex_output_map
            }
        )

        nn.init.xavier_uniform_(self.embedding.weight)
        nn.init.xavier_uniform_(self.fc_out_pos.weight)

        for pos_class in contlex_output_map:
            nn.init.xavier_uniform_(self.fc_out_contlex[str(pos_class)].weight)

        self.pos_weight = 1.0
        self.contlex_weight = 1.0

        self.learning_rate = learning_rate
        self.batch_size = batch_size
        self.contlex_output_map = contlex_output_map
        self.pos_class_sorted = list(sorted(contlex_output_map.keys()))
        self.contlex_class_max = max(contlex_output_map.values())

        self.pos_f1 = MulticlassF1Score(num_classes=pos_num_classes, average="weighted")
        self.contlex_f1 = MulticlassF1Score(num_classes=self.contlex_class_max, average="weighted")

    def forward(self, x, pos_target=None):
        x = self.embedding(x)
        transformer_out = self.transformer_encoder(x)

        transformer_last_hidden = transformer_out[:, -1, :]

        pos_output = self.fc_out_pos(transformer_last_hidden)

        pos_labels = pos_target if pos_target is not None else torch.argmax(pos_output, dim=1) 

        contlex_output = torch.zeros((x.size(0), self.contlex_class_max), device=self.device)

        for pos_class in self.pos_class_sorted:
            contlex_size = self.hparams.contlex_output_map[pos_class]
            indices = (pos_labels == int(pos_class)).nonzero(as_tuple=True)[0]

            if len(indices) > 0:
                fc_out_contlex = self.fc_out_contlex[str(pos_class)]
                contlex_out = fc_out_contlex(transformer_last_hidden[indices])
                contlex_output[indices, :contlex_size] = contlex_out

        return pos_output, contlex_output

    def custom_loss(self, pos_output, contlex_output, pos_target, contlex_target):
        pos_loss = nn.CrossEntropyLoss()(pos_output, pos_target)

        contlex_loss = nn.CrossEntropyLoss()(contlex_output, contlex_target)

        total_loss = self.pos_weight * pos_loss + self.contlex_weight * contlex_loss
        return total_loss, pos_loss, contlex_loss

    def training_step(self, batch, batch_idx):
        x, pos_y, contlex_y = batch
        pos_output, contlex_output = self(x, pos_target=pos_y)

        total_loss, pos_loss, contlex_loss = self.custom_loss(pos_output, contlex_output, pos_y, contlex_y)
        self.log("train_loss", total_loss, on_step=True, on_epoch=True, prog_bar=True)
        self.log("train_pos_loss", pos_loss, on_step=True, on_epoch=True, prog_bar=True)
        self.log("train_contlex_loss", contlex_loss, on_step=True, on_epoch=True, prog_bar=True)

        return total_loss

    def validation_step(self, batch, batch_idx):
        x, pos_y, contlex_y = batch
        pos_output, contlex_output = self(x, pos_target=pos_y)

        total_loss, pos_loss, contlex_loss = self.custom_loss(pos_output, contlex_output, pos_y, contlex_y)
        self.log("val_loss", total_loss, on_step=False, on_epoch=True, prog_bar=True)
        self.log("val_pos_loss", pos_loss, on_step=False, on_epoch=True, prog_bar=True)
        self.log("val_contlex_loss", contlex_loss, on_step=False, on_epoch=True, prog_bar=True)

        pos_preds = torch.argmax(pos_output, dim=1)
        pos_acc = (pos_preds == pos_y).float().mean()
        self.log("val_pos_acc", pos_acc, on_step=False, on_epoch=True, prog_bar=True)

        contlex_preds = torch.argmax(contlex_output, dim=1)                
        contlex_acc = (contlex_preds == contlex_y).float().mean()
        self.log(f"val_contlex_acc", contlex_acc, on_step=False, on_epoch=True, prog_bar=True)

        self.pos_f1(pos_preds, pos_y)
        self.log("val_pos_f1", self.pos_f1, on_step=False, on_epoch=True, prog_bar=True)

        self.contlex_f1(contlex_preds, contlex_y)
        self.log("val_contlex_f1", self.contlex_f1, on_step=False, on_epoch=True, prog_bar=True)

        self.log("lr", self.trainer.optimizers[0].param_groups[0]['lr'], on_step=False, on_epoch=True, prog_bar=True)


    def configure_optimizers(self):
        optimizer = torch.optim.AdamW(self.parameters(), lr=self.learning_rate)
        # scheduler = torch.optim.lr_scheduler.ReduceLROnPlateau(optimizer, mode='min', factor=0.95, patience=10)
        scheduler = torch.optim.lr_scheduler.CosineAnnealingLR(optimizer, T_max=25)
        # scheduler = torch.optim.lr_scheduler.ExponentialLR(optimizer, gamma=0.95)
        return {"optimizer": optimizer, "lr_scheduler": scheduler, "monitor": "val_loss"}
        
    
    def train_dataloader(self):
        return DataLoader(train_dataset, batch_size=self.batch_size, shuffle=True)
    
    def val_dataloader(self):
        return DataLoader(val_dataset, batch_size=self.batch_size, shuffle=True)

# Hyperparameters
embed_size = 128
hidden_size = 512
num_layers = 3
nhead = 8
dropout = 0.2
epochs = 100
learning_rate = 0.003
batch_size = 512

pos_num_classes = len(set(pos_labels))

date_time_str = datetime.now().strftime("%Y-%m-%d-%H-%M-%S")

model = SharedEmbeddingTransformer(
    pos_num_classes, 
    contlex_output_map, 
    vocab_size, 
    embed_size=embed_size,
    hidden_size=hidden_size, 
    num_layers=num_layers,
    nhead=nhead,
    dropout=dropout,
    learning_rate=learning_rate,
    batch_size=batch_size
)

print(model.hparams)

checkpoint_callback = ModelCheckpoint(
     monitor='val_loss',
     dirpath=f"./model/{date_time_str}/",
     filename='model-{epoch:02d}-{val_loss:.2f}-{val_contlex_f1:.2f}',
     mode="min",
     enable_version_counter=True,
    #  save_last=True,
)

early_stop_callback = EarlyStopping(monitor="val_loss", min_delta=0.00, patience=10, verbose=False, mode="min")


trainer = L.Trainer(
    min_epochs=100,
    max_epochs=epochs,
    accumulate_grad_batches=4,
    callbacks=[
        # early_stop_callback, 
        StochasticWeightAveraging(swa_lrs=1e-2), checkpoint_callback]
)

# tuner = Tuner(trainer)
# tuner.scale_batch_size(model, mode="power")
# lr_finder = tuner.lr_find(model)

# fig = lr_finder.plot(suggest=True)
# fig.show()

# # TO TRAIN, uncomment
# trainer.fit(model, train_dataloader, val_dataloader)

model= SharedEmbeddingTransformer.load_from_checkpoint(f"./model/BEST/final_model.ckpt")

test_data

print(model.hparams)
model.eval()
model = model.to(device)

for i, row in test_data.iterrows():
    x_tokenized = sp.encode(row["X"], out_type=int)
    x = torch.tensor([x_tokenized], dtype=torch.long).to(device)
    pos_output, _ = model(x)
    pos_pred = torch.argmax(pos_output, dim=1)

    pos_target_main = encoder_manager.transform_pos([row["pos"]])
    pos_target = torch.tensor(pos_target_main, dtype=torch.long).to(device)

    _, contlex_output = model(x, pos_target=pos_target)
    contlex_pred = torch.argmax(contlex_output, dim=1)
    pos_label = encoder_manager.inverse_transform_pos([pos_pred.item()])[0]
    contlex_label = encoder_manager.inverse_transform_contlex([row["pos"]], [contlex_pred.item()])[0]
    test_data.at[i, f"pos_pred"] = pos_label
    test_data.at[i, f"contlex_pred"] = contlex_label

print(classification_report(test_data["pos"], test_data[f"pos_pred"]))
print(classification_report(test_data["label"], test_data[f"contlex_pred"]))


max_len_of_X_words = test_data["X"].apply(lambda x: x.count(" ")).max()
max_len_of_X_words

for j in range(1, max_len_of_X_words + 1):
    print(f"Max number of words: {j}")

    for i, row in test_data.iterrows():
        X_with_max_j_words = " ".join(row["X"].split()[:j])
        x_tokenized = sp.encode(X_with_max_j_words, out_type=int)
        x = torch.tensor([x_tokenized], dtype=torch.long).to(device)
        pos_output, _ = model(x)
        pos_pred = torch.argmax(pos_output, dim=1)

        pos_target_main = encoder_manager.transform_pos([row["pos"]])
        pos_target = torch.tensor(pos_target_main, dtype=torch.long).to(device)

        _, contlex_output = model(x, pos_target=pos_target)
        contlex_pred = torch.argmax(contlex_output, dim=1)
        pos_label = encoder_manager.inverse_transform_pos([pos_pred.item()])[0]
        contlex_label = encoder_manager.inverse_transform_contlex([row["pos"]], [contlex_pred.item()])[0]
        test_data.at[i, f"pos_pred"] = pos_label
        test_data.at[i, f"contlex_pred"] = contlex_label

    print("POS Accuracy:", accuracy_score(test_data["pos"], test_data[f"pos_pred"]))
    print("Contlex Accuracy:", accuracy_score(test_data["label"], test_data[f"contlex_pred"]))
    print()
